AutoHotKey script to remove the title bar from a window
This script is useful for PowerPoint presentations when presenting in a
conference call where you might not want to have the presentation full
screen (since you then need to share the full screen). Instead, set the
presetnation to "window" mode, start the presentation, hit ctrl+shift+a
and the title bar and status will be removed.

You can now share this window in your conference call and it will look
full screen but not take up as much realestate or eat as many resources.

To move the window when in this mode hold down ctrl+alt+shift and click
and drag anywhere in the window.